+++ dataloom/tests/postgres/test_eager_loading_pg.py
@@ -1088,3 +1088,216 @@ class TestEagerLoadingOnPG:
         conn.close()
 
         conn.close()
+
+    def test_self_relations_eager(self):
+        from dataloom import (
+            Loom,
+            Model,
+            Column,
+            PrimaryKeyColumn,
+            TableColumn,
+            ForeignKeyColumn,
+            ColumnValue,
+            Include,
+        )
+        from dataloom.keys import PgConfig
+
+        pg_loom = Loom(
+            dialect="postgres",
+            database=PgConfig.database,
+            password=PgConfig.password,
+            user=PgConfig.user,
+        )
+
+        class Employee(Model):
+            __tablename__: TableColumn = TableColumn(name="employees")
+            id = PrimaryKeyColumn(type="int", auto_increment=True)
+            name = Column(type="text", nullable=False, default="Bob")
+            supervisorId = ForeignKeyColumn(
+                "Employee", maps_to="1-1", type="int", required=False
+            )
+
+        conn, tables = pg_loom.connect_and_sync([Employee], force=True)
+
+        empId = pg_loom.insert_one(
+            instance=Employee, values=ColumnValue(name="name", value="John Doe")
+        )
+
+        _ = pg_loom.insert_bulk(
+            instance=Employee,
+            values=[
+                [
+                    ColumnValue(name="name", value="Michael Johnson"),
+                    ColumnValue(name="supervisorId", value=empId),
+                ],
+                [
+                    ColumnValue(name="name", value="Jane Smith"),
+                    ColumnValue(name="supervisorId", value=empId),
+                ],
+            ],
+        )
+
+        emp_and_sup = pg_loom.find_by_pk(
+            instance=Employee,
+            pk=2,
+            select=["id", "name", "supervisorId"],
+            include=Include(
+                model=Employee,
+                has="one",
+                select=["id", "name"],
+                alias="supervisor",
+            ),
+        )
+
+        assert emp_and_sup == {
+            "id": 2,
+            "name": "Michael Johnson",
+            "supervisorId": 1,
+            "supervisor": {"id": 1, "name": "John Doe"},
+        }
+        emp_and_sup = pg_loom.find_by_pk(
+            instance=Employee,
+            pk=1,
+            select=["id", "name", "supervisorId"],
+            include=Include(
+                model=Employee,
+                has="one",
+                select=["id", "name"],
+                alias="supervisor",
+            ),
+        )
+        assert emp_and_sup == {
+            "id": 1,
+            "name": "John Doe",
+            "supervisorId": None,
+            "supervisor": None,
+        }
+        conn.close()
+
+    def test_n2n_relations_eager(self):
+        from dataloom import (
+            Loom,
+            Model,
+            Column,
+            PrimaryKeyColumn,
+            TableColumn,
+            ForeignKeyColumn,
+            ColumnValue,
+            Include,
+        )
+        from dataloom.keys import PgConfig
+
+        pg_loom = Loom(
+            dialect="postgres",
+            database=PgConfig.database,
+            password=PgConfig.password,
+            user=PgConfig.user,
+        )
+
+        class Course(Model):
+            __tablename__: TableColumn = TableColumn(name="courses")
+            id = PrimaryKeyColumn(type="int", auto_increment=True)
+            name = Column(type="text", nullable=False, default="Bob")
+
+        class Student(Model):
+            __tablename__: TableColumn = TableColumn(name="students")
+            id = PrimaryKeyColumn(type="int", auto_increment=True)
+            name = Column(type="text", nullable=False, default="Bob")
+
+        class StudentCourses(Model):
+            __tablename__: TableColumn = TableColumn(name="students_courses")
+            studentId = ForeignKeyColumn(table=Student, type="int")
+            courseId = ForeignKeyColumn(table=Course, type="int")
+
+        conn, tables = pg_loom.connect_and_sync(
+            [Student, Course, StudentCourses], force=True
+        )
+
+        # insert the courses
+        mathId = pg_loom.insert_one(
+            instance=Course, values=ColumnValue(name="name", value="Mathematics")
+        )
+        engId = pg_loom.insert_one(
+            instance=Course, values=ColumnValue(name="name", value="English")
+        )
+        phyId = pg_loom.insert_one(
+            instance=Course, values=ColumnValue(name="name", value="Physics")
+        )
+
+        # create students
+
+        stud1 = pg_loom.insert_one(
+            instance=Student, values=ColumnValue(name="name", value="Alice")
+        )
+        stud2 = pg_loom.insert_one(
+            instance=Student, values=ColumnValue(name="name", value="Bob")
+        )
+        stud3 = pg_loom.insert_one(
+            instance=Student, values=ColumnValue(name="name", value="Lisa")
+        )
+
+        # enrolling students
+        pg_loom.insert_bulk(
+            instance=StudentCourses,
+            values=[
+                [
+                    ColumnValue(name="studentId", value=stud1),
+                    ColumnValue(name="courseId", value=mathId),
+                ],  # enrolling Alice to mathematics
+                [
+                    ColumnValue(name="studentId", value=stud1),
+                    ColumnValue(name="courseId", value=phyId),
+                ],  # enrolling Alice to physics
+                [
+                    ColumnValue(name="studentId", value=stud1),
+                    ColumnValue(name="courseId", value=engId),
+                ],  # enrolling Alice to english
+                [
+                    ColumnValue(name="studentId", value=stud2),
+                    ColumnValue(name="courseId", value=engId),
+                ],  # enrolling Bob to english
+                [
+                    ColumnValue(name="studentId", value=stud3),
+                    ColumnValue(name="courseId", value=phyId),
+                ],  # enrolling Lisa to physics
+                [
+                    ColumnValue(name="studentId", value=stud3),
+                    ColumnValue(name="courseId", value=engId),
+                ],  # enrolling Lisa to english
+            ],
+        )
+
+        alice = pg_loom.find_by_pk(
+            Student,
+            pk=stud1,
+            select=["id", "name"],
+            include=Include(model=Course, junction_table=StudentCourses, has="many"),
+        )
+
+        assert alice == {
+            "id": 1,
+            "name": "Alice",
+            "courses": [
+                {"id": 1, "name": "Mathematics"},
+                {"id": 2, "name": "English"},
+                {"id": 3, "name": "Physics"},
+            ],
+        }
+
+        english = pg_loom.find_by_pk(
+            Course,
+            pk=engId,
+            select=["id", "name"],
+            include=Include(model=Student, junction_table=StudentCourses, has="many"),
+        )
+
+        assert english == {
+            "id": 2,
+            "name": "English",
+            "students": [
+                {"id": 1, "name": "Alice"},
+                {"id": 2, "name": "Bob"},
+                {"id": 3, "name": "Lisa"},
+            ],
+        }
+        conn.close()